rename metrics_fetcher module to daqmetrics
