+++ daqmetrics.py
