+++ trex_HV_gui.py
@@ -19,7 +19,7 @@ import utils
 from checkframe import ChecksFrame
 from check import load_checks_from_toml_file
 from utilsgui import PrintToTextWidget, ToolTip, enable_children, validate_numeric_entry_input
-from metrics_fetcher import MetricsFetcherSSH, FeminosDaqMetrics, FemDaqMetrics
+from daqmetrics import MetricsFetcherSSH, FeminosDaqMetrics, FemDaqMetrics
 from daqmetricsgui import DaqMetricsGUI
 import logger
